Updating styling, add new server and remove server commands
