+++ ConnectionManager/SpeckleConnectionManagerUI/.gitignore
@@ -1,0 +1,8 @@
+.idea/
+.vscode/
+.vs/
+
+bin/
+obj/
+
+*.user
