+++ ConnectionManager/SpeckleConnectionManagerUI/App.axaml
@@ -1,0 +1,19 @@
+<Application xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="using:SpeckleConnectionManagerUI"
+             xmlns:themes="clr-namespace:Material.Styles.Themes;assembly=Material.Styles"
+             xmlns:icons="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
+             x:Class="SpeckleConnectionManagerUI.App">
+  <Application.DataTemplates>
+    <local:ViewLocator />
+  </Application.DataTemplates>
+  <Application.Resources>
+    <themes:BundledTheme
+      BaseTheme="Light"
+      PrimaryColor="Blue"
+      SecondaryColor="Amber" />
+  </Application.Resources>
+  <Application.Styles>
+    <StyleInclude Source="avares://Material.Avalonia/Material.Avalonia.Templates.xaml" />
+  </Application.Styles>
+</Application>
