+++ ConnectionManager/SpeckleConnectionManagerUI/Services/Authenticate.cs
