+++ ConnectionManager/SpeckleConnectionManagerUI/Services/BackgroundRefreshTokenProcess.cs
