+++ ConnectionManager/SpeckleConnectionManagerUI/Services/Constants.cs
