+++ ConnectionManager/SpeckleConnectionManagerUI/Services/TimedRefreshToken.cs
