+++ ConnectionManager/SpeckleConnectionManagerUI/ViewModels/ViewModelBase.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using ReactiveUI;
+
+namespace SpeckleConnectionManagerUI.ViewModels
+{
+    public class ViewModelBase : ReactiveObject
+    {
+    }
+}
