+++ ConnectionManager/SpeckleConnectionManagerUI/s2block.ico
