--- SpeckleConnectionManagerUI/.gitignore
@@ -1,8 +1,0 @@
-.idea/
-.vscode/
-.vs/
-
-bin/
-obj/
-
-*.user
