--- SpeckleConnectionManagerUI/App.axaml
@@ -1,12 +1,0 @@
-<Application xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="using:SpeckleConnectionManagerUI"
-             x:Class="SpeckleConnectionManagerUI.App">
-    <Application.DataTemplates>
-        <local:ViewLocator />
-    </Application.DataTemplates>
-
-    <Application.Styles>
-        <FluentTheme Mode="Dark" />
-    </Application.Styles>
-</Application>
