--- SpeckleConnectionManagerUI/ViewLocator.cs
@@ -1,32 +1,0 @@
-using System;
-using Avalonia.Controls;
-using Avalonia.Controls.Templates;
-using SpeckleConnectionManagerUI.ViewModels;
-
-namespace SpeckleConnectionManagerUI
-{
-    public class ViewLocator : IDataTemplate
-    {
-        public bool SupportsRecycling => false;
-
-        public IControl Build(object data)
-        {
-            var name = data.GetType().FullName!.Replace("ViewModel", "View");
-            var type = Type.GetType(name);
-
-            if (type != null)
-            {
-                return (Control)Activator.CreateInstance(type)!;
-            }
-            else
-            {
-                return new TextBlock { Text = "Not Found: " + name };
-            }
-        }
-
-        public bool Match(object data)
-        {
-            return data is ViewModelBase;
-        }
-    }
-}
