--- SpeckleConnectionManagerUI/ViewModels/ViewModelBase.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using ReactiveUI;
-
-namespace SpeckleConnectionManagerUI.ViewModels
-{
-    public class ViewModelBase : ReactiveObject
-    {
-    }
-}
