--- SpeckleConnectionManagerUI/Views/ConnectStatusView.axaml
@@ -1,44 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="SpeckleConnectionManagerUI.Views.ConnectStatusView">
-    <DockPanel>
-        <Button FontSize="20"
-                DockPanel.Dock="Bottom"
-                HorizontalAlignment="Stretch"
-                HorizontalContentAlignment="Center"
-                Command="{Binding $parent[Window].DataContext.DeleteAllAuthData}">
-            Delete All Auth Data
-        </Button>
-        <StackPanel>
-            <TextBlock Foreground="LightSlateGray" Margin="50" Height="100" TextAlignment="Center" FontSize="40">Speckle@Arup Account Manager</TextBlock>
-            <ItemsControl Margin="50 0 0 0" Items="{Binding Items}">
-                <ItemsControl.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel>
-                            <Grid Margin="0 0 0 10" ColumnDefinitions="Auto,Auto" RowDefinitions="Auto,Auto">
-                                <Button FontSize="20"
-                                        Grid.Row="0"
-                                        Grid.Column="0"
-                                        IsEnabled="{Binding Disconnected}"
-                                        Command="{Binding $parent[Window].DataContext.ConnectToServer}"
-                                        CommandParameter="{Binding Identifier}">
-                                    Connect
-                                </Button>
-                                <TextBlock VerticalAlignment="Center"
-                                           Margin="5 0 0 0"
-                                           FontSize="20"
-                                           Grid.Row="0"
-                                           Grid.Column="1"
-                                           Text="{Binding ServerName}"
-                                           Foreground="{Binding Colour}" />
-                            </Grid>
-                        </StackPanel>
-                    </DataTemplate>
-                </ItemsControl.ItemTemplate>
-            </ItemsControl>
-        </StackPanel>
-    </DockPanel>
-</UserControl>
