--- SpeckleConnectionManagerUI/Views/MainWindow.axaml
@@ -1,12 +1,0 @@
-<Window xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="SpeckleConnectionManagerUI.Views.MainWindow"
-        Icon="/Assets/s2block.ico"
-        Title="Speckle@Arup Account Manager"
-        Background="Transparent"
-        Width="800" Height="500"
-        Content="{Binding Content}">
-</Window>
